requirements.txt
pillow
transformers
torch
osxphotos
ultralytics
cv2
matplotlib